1132 remove custom application event listeners outside atlas engine
